requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
SQLAlchemy==2.0.45
psycopg2-binary==2.9.11
python-multipart==0.0.21
python-dotenv==1.2.1
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
